requirements.txt
flask>=2.0
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
joblib
matplotlib
seaborn
jupyter
notebook
ipykernel
python-dotenv>=0.21.0
gunicorn>=20.1.0
